fix: alignment of 0% label on slider (#247)
* fix: alignment of 0% label on slider

* fix: don't set right position of slider button unless necessary

The right position was being set to 0 when the left position was being used
to position the button, which causes the element to be wider than necessary
and may have been causing display issues on certain browsers.

Also sets transform to "none" unless needed. That's the default value.

* refactor: fix typo in variable name

* refactor: combine css styles into single switch/case

* refactor: don't set values for unused css properties

Before, we set a default value that matched the browser's,
but there's no need to do that. Simply leaving them undefined will cause
them to not be set in the first place.