fix: move simplified version of react-terra's useTx hook into common module
This fixes ASTRO-1218 because the useTx hook no longer tracks the tx after broadcast.

The useTx ends at onBroadcasting or onError. In the former scenario (a successful post),
the caller is now responsible for tracking the tx (e.g. with useTxInfo).

For that reason, the useTx hook no longer accepts an onSuccess callback.

The new useTx hook differs from react-terra's in the following ways:
- onSuccess has been removed
- txHash state is no longer managed by the hook
- onError is passed two parameters:
  1. a TS enum that maps to the original error type
  2. The original, unmodified, Error from useWallet's post function
     (to get additional context, if needed)