fix: explicitly pass props to SlippagePopover to elliminate ref warning
The Controller's render function includes a ref in the field props,
so passing all field props via a spread passes that ref to the component,
which causes React to complain, even though the ref isn't used.

This removes the spread in favor if explicitly passing the two
props that are used: value and onChange.